requirements.txt
beautifulsoup4
lxml
requests
pytz
tzdata
python-telegram-bot
groq
dotenv
uvicorn
fastapi